Use BigInteger in place of int for ring conversion & exponents
Replace 'Ring::fromInt' with 'Ring::from(BigInteger|Long)' for later
support of large prime fields, such as those arising during elliptic
curve cryptography. Similarly, make 'Ring::power' take a BigInteger or
long exponent instead of an int and make 'FiniteField::size' return a
BigInteger (or a saturated long using 'FiniteField::sizeAsLong').

Also add a default 'AbelianGroup::product' method for multiplication by
BigInteger (or long) scalars, using the same generic exponentiation
algorithm as that used in 'Ring::power'. This will later be useful for
scalar multiplication in elliptic curve and Diffie-Hellman groups
(modulo the caveat that the default method isn't constant-time).

+++ src/test/java/securecompute/helper/WithDefaultField.java
@@ -2,6 +2,8 @@ package securecompute.helper;
 
 import securecompute.algebra.Field;
 
+import java.math.BigInteger;
+
 public interface WithDefaultField<E> extends WithDefaultRing<E>, Field<E> {
 
     @Override
@@ -23,7 +25,7 @@ public interface WithDefaultField<E> extends WithDefaultRing<E>, Field<E> {
     }
 
     @Override
-    default E power(E elt, int exponent) {
+    default E power(E elt, BigInteger exponent) {
         return getDefaultStructure().power(elt, exponent);
     }
 }
